[VAN] compact source, readjust ui star

+++ app/src/main/java/com/example/balo/adapter/product/AdminProductEditAdapter.kt
@@ -22,7 +22,7 @@ class AdminProductEditAdapter(
                 Utils.displayBase64Image(item.pic, imgPic)
                 tvName.text = item.name
                 tvPrice.text = item.priceSell
-                val available = item.quantitiy.toInt() - item.sell.toInt()
+                val available = item.quantitiy.toFloat() - item.sell.toFloat()
                 tvQuantity.text = available.toString()
                 cbDelete.setOnClickListener {
                     onCheckBox.invoke(Pair(cbDelete.isChecked, item.id))
